ci: scan alternate clang-tidy configurations

+++ wirelog/columnar/relation.c
@@ -1971,8 +1971,9 @@ radix_sort_k16(col_rel_t *r, uint32_t start_row, uint32_t nrows)
             "uniform_+_count=%.0f%% scatter=%.0f%% apply=%.0f%% "
             "total_ms=%.3f\n",
             nrows, nc, _nPs, _nSk,
-            100.0 * _tU / (_tot + 1),
-            100.0 * _tS / (_tot + 1), 100.0 * _tA / (_tot + 1),
+            100.0 * (double)_tU / (double)(_tot + 1),
+            100.0 * (double)_tS / (double)(_tot + 1),
+            100.0 * (double)_tA / (double)(_tot + 1),
             (double)_tot * 1e-6);
     }
 #endif
@@ -2149,8 +2150,9 @@ col_rel_radix_sort(col_rel_t *r, uint32_t start_row, uint32_t nrows)
             "uniform_+_count=%.0f%% scatter=%.0f%% apply=%.0f%% "
             "total_ms=%.3f\n",
             nrows, nc, _nPs, _nSk,
-            100.0 * _tU / (_tot + 1),
-            100.0 * _tS / (_tot + 1), 100.0 * _tA / (_tot + 1),
+            100.0 * (double)_tU / (double)(_tot + 1),
+            100.0 * (double)_tS / (double)(_tot + 1),
+            100.0 * (double)_tA / (double)(_tot + 1),
             (double)_tot * 1e-6);
     }
 #endif
